Add tables and database merging to IndexDb.
A table has a list of column names.  A column name can be blank, but if
it isn't, it corresponds to the string table of the same name.  Each row in
the table is a tuple of integers.  The integer value for a cell with a
named column is an ID for the associated string table, and the integer
value will be appropriately remapped during database merging.

If a column name is blank, then no remapping occurs for that column's
integers.  This is useful for lines and columns.

Each integer is encoded using a variant of LLVM's variable-length-encoding.
Instead of using 4-bit nibbles, however, I use 8-bit bytes.  (i.e. The
first byte has the lowest 7 bits of value, and the highest bit is set iff
there is a second byte.)

The rows are encoded as variable-length, NUL-terminated strings.  There is
no table-of-contents, but the strings are stored in sorted order, so a
binary search should be possible.  That isn't implemented just yet, but
will be soon.