Add GotoWindow, accessible with Ctrl-S.
The window shows all of the index's symbol USRs.  The window has a search
box that filters the symbol list as the user types.  Keyboard focus is
always on the search box, not on the results list, but vertical navigation
keys control the results list anyway (i.e. Up/Down/PgUp/PgDn).  Home and
End move the search box cursor, so Ctrl-Home and Ctrl-End affect the
results list.

The results list is filtered using the RE2 library, which is currently an
external dependency at /home/rprichard.  I'll probably include its source
inside this repository.

Despite using RE2, the search box does not currently accept a regex.
Instead, the search query is split into words comprised of identifier
characters [A-Za-z0-9_] and '*' glob characters.  The GotoWindow computes
a regex using these rules:
 - Each word must begin and end on a word boundary.
 - '*' matches 0-or-more identifier characters.
 - The last word in the query must end with the end of the USR.

These rules will likely need updating -- I'd like to replace USRs with
something more user-friendly, and currently it appears impossible to match
a C++ function because the USR ends with '#'.  (It *is* possible to match
a parameter/variable inside a C++ function.  In any case, C++ functions are
troublesome due to parameter types appearing in the USR.)

The filtering is run in the background, parallelized, using QFuture and
QFutureWatcher.  I'm not convinced that my cleanup/deallocation code for
QFutureWatcher is correct.