Checkpoint clang-indexer table-sorting work.
indexdb is currently broken -- the rows in the non-string tables are
logically sorted but not lexicographically sorted.  The Table::lowerBound
routines uses lexicographic comparisons, so lookups don't work.  I'm
planning to change Table::lowerBound to do logical comparisons instead.

Explicitly use little-endian encoding in the file.  Maybe I should use
big-endian instead because that's the official "network byte order," but
it seems that all the popular hardware (x86/ARM) is little-endian.

Always use a prime number for the size of the string tables.  This is a
cheap way to reduce hash collisions a bit.