Determine the source locations for an #include filename.
This task is made more difficult because the Clang 3.1
clang::PPCallbacks::InclusionDirective does not provide the location of the
start of the filename -- it only provides endLoc, which is the beginning of
the last token of the filename.  It looks like Clang HEAD currently
provides a clang::SourceRange instead of endLoc.

In general, an angle-bracketed filename can be expanded from a macro, so
there isn't one simple range of source file characters to associate with
the inclusion, but this code will at least highlight one character, and it
should usually do something sensible.