When building an archive, align each entry to 8 bytes.
This ensures that aligned items within each entry are aligned to both the
start of the file and the start of the entry, which is useful if an
individual entry is mapped into memory, or (hypothetically) if an entry
were extracted and written to a new file.

Technically, alignment relative to the file is the only thing that matters
for memory-mapping because mmap only works with page-aligned offsets and
addresses.  However, I have an abstraction in a later commit that can map
an arbitrary non-page-aligned offset into memory, and it's useful to be
able to align items relative to the view.