Add a new TableReport class and associated view/window classes.
TableReportView is similar to QTreeView, but also includes functionality
from GotoWindow.  It shows a list of rows with a fixed number of columns.
The data can be sorted by one of the columns or filtered using a regex.
Sorting is done synchronously (i.e. the GUI thread is blocked), but
filtering is performed in the background using QtConcurrent::mappedReduced.

Sorting and filtering are managed using index->index maps kept in instances
of TableReportView_ProxyReport subclasses.  A ProxyReport object can map
an index to and from its source (which can be the TableReport or another
ProxyReport).  (Mapping an index from the source to the proxy is needed for
tracking a selected item when the sort/filter maps change.)

The {Proxy,Table}Report objects look like this once the filtering is
complete:

 If sorted:
  FilterProxyReport -> SortProxyReport -> DirectProxyReport -> TableReport

 Otherwise:
  FilterProxyReport ->                    DirectProxyReport -> TableReport

I initially tried to do both sorting and filtering on a background thread.
There were problems:
 - The sorting is slow.  (IIRC, it was something like 200ms to sort the
   global-defs data for LLVM+Clang.)
 - It's hard to cancel the sorting.  std::sort doesn't accept an atomic
   cancel flag that another thread could set.  If canceling is necessary,
   it would apparently require a custom sort routine.

TableReportWindow is analogous to TreeReportWindow, but it also includes an
optional filter box (hidden by default).