Redo the internal dependency handling.  Stop using prl files.
Apparently, the MinGW makefile generator does not read prl files for
libraries listed on the LIBS variable.[1]
[1] http://www.archivum.info/qt-interest@trolltech.com/2009-06/00888/%28Qt-interest%29-qmake-prl-files-on-MinGW-not-working-as-expected.html

The previous project structure was also causing some kind of error with
"qmake -r" on Linux.  For example:
    Failure to open file: /home/rprichard/sourceweb-build/clang-indexer-qmake/../clang-indexer/Makefile
The error goes away when I rerun "qmake -r".  Maybe the problem is that a
qmake "subdir" is not supposed to exist in a sibling directory.  It goes
away with the new code for handling dependencies.

Each directory has a dependencies.cfg file.  The add_dependencies.pri
script parses the file for a project and adds direct dependencies to the
include path.  It looks for indirect dependencies and adds them to the
LIBS variables.

It would have been natural to use recursion in the addDependencies
function, but when I attempted that, something strange happened involving
variables.  At one point, I noticed that when the second instance of a
recursive function returned to the first instance, all of the variables
(local, global, exported, etc) were apparently cleared.  A variable could
be set and read within this new blank environment, but exporting it had no
effect.  The outermost context could read the value of a variable set in
the second instance, but the first instance would see a blank variable.  My
current guess is that I've hit a qmake bug.