Win32 portability: implement a Win32 version of indexer::Process.
Unfortunately, this code is untested.  It builds, but sw-clang-indexer
crashes before reaching main().  I think there is an incompatibility
between the prebuilt Qt 4.8.4 binaries and MinGW G++ 4.7.2.

The pipes to/from the daemon process are opened in text mode.  I'm not
entirely sure this is right, but I think the daemon will open its own
pipes in text mode, too.