Add a configure script.
It's now possible to build and install SourceWeb using a conventional set
of commands:

    [in any directory, including the root of the source directory]
    <srcdir>/configure --prefix=<prefix> --with-clang-dir=<clang-dir>
    make [-j<nproc>]
    make install

The typical qmake build process also still works:

    [in any directory, including the root of the source directory]
    qmake [-r] <srcdir>/sourceweb.pro CLANG_DIR=<clang-dir>
    make [-j<nproc>]
    make install INSTALL_ROOT=<prefix>

I like keeping the configure script optional because I want the ability to
open the project in the QtCreator IDE.  The IDE will write a .pro.user file
into the source directory, but it otherwise leaves the source directory
untouched.  It runs qmake and make in separate build directories (typically
one for Debug and one for Release).

In the configure script and qmake files, harmonize the use of the name
Clang over the name LLVM.

Clean out obsolete things from .gitignore.