OS X portability:  Make C++11 compilation work.
On OS X, there are two standard C++ libraries: libstdc++ and libc++.
The libstdc++ is old.  (I think it came from G++ 4.2.)  It does not
support C++11.  The libc++ is a newer implementation that's part of the
LLVM project.  The default library is libstdc++, but libc++ can be
selected using -stdlib=libc++.  Therefore, create a enable-cxx11.pri
file that passes -stdlib=libc++ on OS X.

The -stdlib=libc++ option needs to be passed both at compile-time (for
headers) and at link-time.

As I understand it, there is *some* level of binary compatibility
between libc++ and libstdc++ on OS X, but I got a bunch of unresolved
symbol linker errors trying to build sw-clang-indexer, so I've decided
that every C++ library should be built with the same standard library.
Therefore, I'm building all of the third party components with
-stdlib=libc++.  It was convenient to accomplish this by including
enable-cxx11.pri in those component's pro files, which also sets the
same C++ language dialect.

There is a related change to libre2 -- there is a util.h file that
tries to include either <tr1/unordered_set> or just <unordered_set>.
By default, it tries to include <tr1/unordered_set>.  This header is
not provided with either libstdc++ or libc++ on OS X, so we need to
use unordered_set instead.  I could have defined USE_CXX0X, but because
libre2's util.h is included by prog.h, which is included in the
navigator code, I would also have to define USE_CXX0X in the navigator,
and I don't want to do that.  So instead, I hacked re2 to use the c++11
<unordered_set> include.  (I also considered making it look at the
value of __cplusplus, which some compilers set to 201103L for C++11
mode, but G++4.6 sets it to 1.)