When jumping to an identifier not visible on the screen, center it.
It seems to me that the navigator ought to do better.  For example, when
navigating to a function definition, instead of centering the identifier on
the first line of the definition, it ought to center the entire definition.
Or, if the definition does not fit, then it should place the start of the
definition at the top of the viewport.  Ideally, it would also consider the
function's comment preceding the prototype to be part of the function.

I think this change is an improvement, though.