Migrate clang-indexer from Clang 3.1 to Clang 3.2.
This required two changes:

 * Change the clang::PPCallbacks::InclusionDirective override to handle
   the new way the filename location is conveyed.  The new method fixes the
   exact problem I was working around, so the new code is much simpler.

   The Clang-reported locations for the start and end of a bracketed
   filename can be on different lines, which results in an Included
   reference with an empty range (same start+end column).  Mitigate this
   problem by adding one to the endColumn for all empty references.

 * Add new libraries to check-clang.pri.

Note that clang-indexer no longer works with Clang 3.1.  I considered
using ifdefs to keep it working with both versions, but decided against it
on the basis that it would increase the maintenance burden too much.  The
migration from 3.1 to 3.2 was fairly trivial, though, so if future
migrations are similarly trivial, then perhaps it could make sense to
release an indexer that works with many versions.  (Ideally, it might make
sense for the indexer to operate as a plugin to eliminate any differences
between the build and index toolchains, in which case supporting many Clang
versions would be desirable.)