Clang 3.2 build support: turn off the -Winitialized warning on Qt code.
This warning breaks with Qt 4.8.1, which is the current version in Ubuntu
12.04.  It's fixed in Qt 4.8.3.

In file included from /usr/include/qt4/QtCore/qsharedpointer.h:50:
/usr/include/qt4/QtCore/qsharedpointer_impl.h:595:43: warning: field 'd' is
      uninitialized when used here [-Wuninitialized]
    inline QWeakPointer(X *ptr) : d(ptr ? d->getAndRef(ptr) : 0), value(ptr)
                                          ^

The problem is that some Qt code initializes a "d" member variable using an
expression like "d->getAndRef(ptr)", where getAndRef is a static method.
Technically, this initialization accesses "d" and invokes undefined
behavior, but because it's a static method call, the "d" is unneeded, so
the code tends to work.

AIUI, because d->getAndRef(ptr) invokes undefined behavior, a clever
compiler could decide that the code must not be reachable and optimize it
away.  Hopefully compilers will not be clever on this Qt code.  Perhaps the
read of "d" disappears before the compiler gets around to doing analysis
like this.

The warning is logged as a Clang bug report, and it's possible that Clang's
behavior will change in the future.
 - http://llvm.org/bugs/show_bug.cgi?id=13127

It appears that Qt fixed this issue:
 - Qt5:
    - http://qt.gitorious.org/qt/qtbase/commit/88cf9402e336fddeb673c92f3c14da47a9f8450b
    - Change-Id: I37d2c4ea54febd40410ca473c906bcb1c66c4974
    - https://codereview.qt-project.org/#change,28837
 - Qt4:
    - http://qt.gitorious.org/qt/qt/commit/0e595ad22db04141bfbd6093824cdcc775fd371d
    - Change-Id: I02a9a1b73f7c18e8342aafeae412759f78086fa6
    - https://codereview.qt-project.org/#change,28838