Use smart pointers, fix a subtle lifetime error in indexTranslationUnit.
The LLVM documentation for ToolInvocation::ToolInvocation is very
misleading regarding the FileManager parameter (probably out-of-date).  It
states:

    /// \param Files The FileManager used for the execution. Class does not take
    /// ownership.

This is technically true, but in fact, the FileManager is intrusively
refcounted.  The caller must create an IntrusiveRefCntPtr for the
FileManager object before passing the pointer forward.  If it doesn't,
it will pass an object with 0-refcount, which can easily result in a
double-free or a memory leak.

If the developer reading that comment thinks it's their responsibility to
free the FileManager, then the code almost always crashes.

AFAIK, this same problem affected clang-indexer when it used 3.2; I'm not
sure why it's just now becoming an issue.