Require that the user enter an index path.
It's better to be explicit here, I think.

Motivation: on OS X, if the user double-clicks the sourceweb.app bundle,
they're likely to get an error, "Index file `index` does not exist."  For
first-time users, this is going to be unfriendly.  It's better to print a
usage message, which will make it clear that running from the command-line
is currently required.